requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
jinja2==3.1.4
gpiozero==1.6.2
pyserial==3.5
pydantic==1.10.18
